Resume/synch check - bypass swap partition UUID check
The timing of when the resume file is created, and the mkswap command
is run, means that this check will always fail unless the --shared-swap
option is used. Bypass check for now - the check for a valid swap is
still in place, so there is no risk introduced. This can be replaced
later if required.

Bump version